Moving kvcache code to a separate class from attentionop
